refactor: date 파일 내용 time으로 이동
