+++ frontend/src/api/journal.ts
@@ -2,7 +2,7 @@ import { httpClient } from '@/api/http';
 import { DogAvatar } from '@/models/dog';
 import { Journal } from '@/models/journal';
 import { Position as NumberPosition, Position } from '@/models/location';
-import { formDate } from '@/utils/date';
+import { formDate } from '@/utils/time';
 
 export const fetchJournals = async (dogId: number, date: string = formDate(new Date())): Promise<Journal[]> => {
     const { data } = await httpClient.get(`/journals?dogId=${dogId}&date=${date}`);
