+++ frontend/src/hooks/useJournals.ts
@@ -1,7 +1,7 @@
 import { fetchJournals } from '@/api/journal';
 import { queryStringKeys } from '@/constants';
 import { Journal } from '@/models/journal';
-import { formDate } from '@/utils/date';
+import { formDate } from '@/utils/time';
 import { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
